23028, add production options to install script

+++ scripts/install_bi.sh
@@ -22,10 +22,12 @@ PRIVATE_DIR=$RUN_DIR/../../private-extensions
 XT_DIR=$RUN_DIR/..
 export BISERVER_HOME=$RUN_DIR/../../ErpBI
 DATABASE=dev
+DATABASEHOST=localhost
+DATABASEPORT=5432
 TENANT=default
 COMMONNAME=$(hostname)
 
-while getopts ":iebcpd:t:n:j:" opt; do
+while getopts ":iebcpd:t:n:j:z:h:o:f:" opt; do
   case $opt in
     e)
       # Install ErpBI and configure
@@ -50,23 +52,35 @@ while getopts ":iebcpd:t:n:j:" opt; do
     d)
       # Set database name to extract
       DATABASE=$OPTARG
-	  echo $DATABASE
       ;;
+    o)
+      # Set database port to extract
+      DATABASEPORT=$OPTARG
+      ;;
+    h)
+      # Set database host to extract
+      DATABASEHOST=$OPTARG
+      ;;	  
     t)
       # Set tenant name
       TENANT=$OPTARG
-	  echo $TENANT
       ;;
     n)
       # Common name for self signed SSL certificate
       COMMONNAME=$OPTARG
-	  echo $COMMONNAME
+      ;;
+    f)
+      # Path for config file
+      CONFIGPATH=$OPTARG
       ;;
     j)
       # Java home
       export JAVA_HOME=$OPTARG
-	  echo $JAVA_HOME
-      ;; 
+      ;;
+    z)
+      # ErpBI.zip path
+      export ERPBIPATH=$OPTARG
+      ;;  
     \?)
       log "Invalid option: -"$OPTARG
       exit 1
@@ -90,21 +104,45 @@ if  ! test -d $BI_DIR ;
 then
 	log ""
 	log "#############################################################"
-	log "#############################################################"
     log "Sorry bi folder not found.  You must clone xtuple/bi"
-	log "#############################################################"
 	log "#############################################################"
 	log ""
     exit 1
 fi
 
-if  [ ! test -d $PRIVATE_DIR && $RUNALL ]  
+if  [ $ERPBIPATH ]
+then
+	if   ! test -f $ERPBIPATH  
+	then
+		log ""
+		log "####################################################################################"
+		log "Sorry can't find ErpBI.zip at "$ERPBIPATH
+		log "####################################################################################"
+		log ""
+		exit 1
+	fi
+fi
+
+if  [ $CONFIGPATH ]
+then
+	if   ! test -f $CONFIGPATH  
+	then
+		log ""
+		log "####################################################################################"
+		log "Sorry can't find config file at "$CONFIGPATH
+		log "####################################################################################"
+		log ""
+		exit 1
+	fi
+else
+	CONFIGPATH=config.js
+fi
+
+if   ! test -d $PRIVATE_DIR && $RUNALL   
 then
 	log ""
 	log "####################################################################################"
-	log "####################################################################################"
     log "Sorry private-extensions folder not found.  You must clone xtuple/private-extensions"
-	log "####################################################################################"
 	log "####################################################################################"
 	log ""
     exit 1
@@ -113,9 +151,7 @@ fi
 install_packages () {
 	log ""
 	log "######################################################"
-	log "######################################################"
 	log "Install prereqs."
-	log "######################################################"
 	log "######################################################"
 	log ""
 	apt-get install -qy git openjdk-6-jdk maven2
@@ -124,9 +160,7 @@ install_packages () {
 	then
 		log ""
 		log "#############################################################"
-		log "#############################################################"
 		log "Sorry can not find javac.  Set Java Home with the -j argument"
-		log "#############################################################"
 		log "#############################################################"
 		log ""
 		exit 1
@@ -136,19 +170,33 @@ install_packages () {
 download_files () {
 	log ""
 	log "######################################################"
-	log "######################################################"
 	log "Download ErpBI, set permissions and generate keystore "
 	log "and truststore for SSL with self signed cert using    "
 	log "common name "$COMMONNAME
 	log "######################################################"
-	log "######################################################"
 	log ""
     cdir $RUN_DIR/../..
-	rm -R ErpBI
-	rm ErpBI.zip
-	wget http://sourceforge.net/projects/erpbi/files/candidate-release/ErpBI.zip/download -O ErpBI.zip
-	unzip ErpBI.zip
-	
+
+	rm -R ErpBI	
+	if  [ $ERPBIPATH ]
+	then
+		log ""
+		log "######################################################"
+		log "Unzipping "$ERPBIPATH
+		log "######################################################"
+		log ""
+		unzip -q $ERPBIPATH
+	else
+		rm ErpBI.zip
+		wget http://sourceforge.net/projects/erpbi/files/candidate-release/ErpBI.zip/download -O ErpBI.zip
+		log ""
+		log "######################################################"
+		log "Unzipping ErpBI.zip"
+		log "######################################################"
+		log ""
+		unzip -q ErpBI.zip
+	fi		
+
 	cdir $BISERVER_HOME/biserver-ce/
 	chmod 755 -R . 2>&1 | tee -a $LOG_FILE
 	
@@ -171,10 +219,8 @@ download_files () {
 run_scripts() {
 	log ""
 	log "######################################################"
-	log "######################################################"
 	log "Build BI solution and move to ErpBI at:               "
 	log $BISERVER_HOME
-	log "######################################################"
 	log "######################################################"
 	log ""
 	cdir $BI_DIR/olap-schema
@@ -203,10 +249,8 @@ run_scripts() {
 configure_pentaho() {
 	log ""
 	log "######################################################"
-	log "######################################################"
 	log "Create datamart database erpbi.  Extract data from dev"
 	log "and load data into tenant default.dev".
-	log "######################################################"
 	log "######################################################"
 	log ""
 	createdb -U postgres -O admin erpbi 2>&1 | tee -a $LOG_FILE
@@ -215,7 +259,7 @@ configure_pentaho() {
 	
 	mv $KETTLE_HOME/.kettle/kettle.properties $KETTLE_HOME/.kettle/kettle.properties.sample  2>&1 | tee -a $LOG_FILE
 	cat $KETTLE_HOME/.kettle/kettle.properties.sample | \
-	sed s'#erpi.source.url=.*#erpi.source.url=jdbc\:postgresql\://localhost\:5432/'$DATABASE'#' | \
+	sed s'#erpi.source.url=.*#erpi.source.url=jdbc\:postgresql\://'$DATABASEHOST'\:'$DATABASEPORT'/'$DATABASE'#' | \
 	sed s'#erpi.tenant.id=.*#erpi.tenant.id='$TENANT'.'$DATABASE'#' \
 	> $KETTLE_HOME/.kettle/kettle.properties  2>&1 | tee -a $LOG_FILE
 	
@@ -225,11 +269,9 @@ configure_pentaho() {
 prep_mobile() {
 	log ""
 	log "######################################################"
-	log "######################################################"
 	log "Prepare mobile app to use the BI Server. Create keys  "
     log "for REST api used by single sign on.  Update config.js"
     log "with BI Server URL https://"$COMMONNAME":8443"
-	log "######################################################"
 	log "######################################################"
 	log ""
 	if [ ! -d $XT_DIR/node-datasource/lib/rest-keys ]
@@ -249,13 +291,12 @@ prep_mobile() {
 	#	sed 'N;s#biServer: {\n        hostname:.*#biServer: {\n        hostname: \"'$COMMONNAME'\",#' \
 	
 	cdir $XT_DIR/node-datasource
-	mv config.js config.js.old 2>&1 | tee -a $LOG_FILE
-	cat config.js.old | \
+	mv $CONFIGPATH $CONFIGPATH'.old' 2>&1 | tee -a $LOG_FILE
+	cat $CONFIGPATH'.old' | \
 	sed 's#restkeyfile: .*#restkeyfile: \"./lib/rest-keys/server.key\",#' | \
 	sed 's#tenantname: .*#tenantname: \"'$TENANT'",#' | \
-	sed 's#bihost: .*#bihost: \"'$COMMONNAME'\",#'| \
-	sed 's#printhost: .*#printhost: \"'$COMMONNAME'\",#' \
-	> config.js
+	sed 's#bihost: .*#bihost: \"'$COMMONNAME'\",#' \
+	> $CONFIGPATH
 }
 
 install_packages
@@ -287,11 +328,9 @@ fi
 
 log ""
 log "######################################################"
-log "######################################################"
 log "                FINISED! READ ME                      "
 log "If you use the self signed certificate created by this"
 log "script you will need to accept the certificate in your"
 log "browser.  Connect to https://"$COMMONNAME":8443"
-log "######################################################"
 log "######################################################"
 log ""
